增加工具类

+++ MuZhiSheQu/3th/zhifu/Alipay/openssl/evp.h
@@ -60,19 +60,19 @@
 #define HEADER_ENVELOPE_H
 
 #ifdef OPENSSL_ALGORITHM_DEFINES
-# include "opensslconf.h"
+# include <openssl/opensslconf.h>
 #else
 # define OPENSSL_ALGORITHM_DEFINES
-# include "opensslconf.h"
+# include <openssl/opensslconf.h>
 # undef OPENSSL_ALGORITHM_DEFINES
 #endif
 
-#include "ossl_typ.h"
+#include <openssl/ossl_typ.h>
 
-#include "symhacks.h"
+#include <openssl/symhacks.h>
 
 #ifndef OPENSSL_NO_BIO
-#include "bio.h"
+#include <openssl/bio.h>
 #endif
 
 /*
@@ -91,7 +91,7 @@
 /* Default PKCS#5 iteration count */
 #define PKCS5_DEFAULT_ITER		2048
 
-#include "objects.h"
+#include <openssl/objects.h>
 
 #define EVP_PK_RSA	0x0001
 #define EVP_PK_DSA	0x0002
@@ -636,14 +636,14 @@ void	EVP_EncodeInit(EVP_ENCODE_CTX *ctx);
 void	EVP_EncodeUpdate(EVP_ENCODE_CTX *ctx,unsigned char *out,int *outl,
 		const unsigned char *in,int inl);
 void	EVP_EncodeFinal(EVP_ENCODE_CTX *ctx,unsigned char *out,int *outl);
-NSInteger	EVP_EncodeBlock(unsigned char *t, const unsigned char *f, NSInteger n);
+int	EVP_EncodeBlock(unsigned char *t, const unsigned char *f, int n);
 
 void	EVP_DecodeInit(EVP_ENCODE_CTX *ctx);
 int	EVP_DecodeUpdate(EVP_ENCODE_CTX *ctx,unsigned char *out,int *outl,
 		const unsigned char *in, int inl);
 int	EVP_DecodeFinal(EVP_ENCODE_CTX *ctx, unsigned
 		char *out, int *outl);
-NSInteger	EVP_DecodeBlock(unsigned char *t, const unsigned char *f, NSInteger n);
+int	EVP_DecodeBlock(unsigned char *t, const unsigned char *f, int n);
 
 void EVP_CIPHER_CTX_init(EVP_CIPHER_CTX *a);
 int EVP_CIPHER_CTX_cleanup(EVP_CIPHER_CTX *a);
